+++ BibleStudyTool.Core/Entities/JoinEntities/TagNote/TagNote.cs
@@ -5,17 +5,17 @@ namespace BibleStudyTool.Core.Entities.JoinEntities
 {
     public class TagNote : BaseEntity
     {
-        public long TagId { get; private set; }
+        public int TagId { get; private set; }
         public Tag Tag { get; }
 
-        public long NoteId { get; private set; }
+        public int NoteId { get; private set; }
         public Note Note { get;}
 
         public TagNote()
         {
         }
 
-        public TagNote(long tagId, long noteId)
+        public TagNote(int tagId, int noteId)
         {
             TagId = tagId;
             NoteId = noteId;
